LayoutTests/platform/mac/accessibility/webarea-size-equals-content-size.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<img src="resources/cake.png" width=1000 height=1000><br>
text

<img src="resources/cake.png" width=1000 height=1000><br>
text

<img src="resources/cake.png" width=1000 height=1000><br>
text

<img src="resources/cake.png" width=1000 height=1000><br>
text

<img src="resources/cake.png" width=1000 height=1000><br>
text

<img src="resources/cake.png" width=1000 height=1000><br>
text

<img src="resources/cake.png" width=1000 height=1000><br>
text

<img src="resources/cake.png" width=1000 height=1000><br>
text

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that an ARIA menu bar is exposed as a menu bar.");

    if (window.accessibilityController) {

        document.getElementById("body").focus();
        shouldBe("accessibilityController.focusedElement.role", "'AXRole: AXWebArea'");
        shouldBe("accessibilityController.focusedElement.parentElement().role", "'AXRole: AXScrollArea'");

        var webAreaHeight = accessibilityController.focusedElement.height;
        var webAreaWidth = accessibilityController.focusedElement.width;

        var scrollHeight = accessibilityController.focusedElement.parentElement().height;
        var scrollWidth = accessibilityController.focusedElement.parentElement().width;

        shouldBeTrue("webAreaWidth > scrollWidth");
        shouldBeTrue("webAreaHeight > scrollHeight");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
